consolidating loadImage methods
